+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.7.0 - 2025-11-07
+
+- Remove non accessible keyboard shortcuts for Accordion component
+- Add new boolean setting `closeOnBlur` for Dropdown component. If `true`, the expanded list will be closed on button blur.
+
 ## 1.6.2 - 2024-12-17
 
 - Add new event `onInit` option for Accordion component.